Separate transformer from helping xalan extension methods
